refactor(folder): entity, dto, controller 폴더 생성 및 패키지 정리
