+++ src/main/java/lotto/dto/LottoResult.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.dto;
 
 import java.util.List;
