+++ src/main/java/lotto/dto/MatchingCountResult.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.dto;
 
 import lotto.constant.WinningCondition;
