+++ src/main/java/lotto/service/LottoResultService.java
@@ -3,9 +3,9 @@ package lotto.service;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
-import lotto.Lotto;
-import lotto.LottoResult;
-import lotto.MatchingCountResult;
+import lotto.dto.LottoResult;
+import lotto.dto.MatchingCountResult;
+import lotto.model.Lotto;
 
 public class LottoResultService {
     private final LottoMatchingCounter lottoMatchingCounter;
