+++ src/test/java/lotto/service/LottoRateCalculatorTest.java
@@ -3,8 +3,8 @@ package lotto.service;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.List;
-import lotto.MatchingCountResult;
 import lotto.constant.WinningCondition;
+import lotto.dto.MatchingCountResult;
 import org.junit.jupiter.api.Test;
 
 class LottoRateCalculatorTest {
