+++ src/test/java/lotto/service/LottoResultServiceTest.java
@@ -6,8 +6,8 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.ArrayList;
 import java.util.List;
-import lotto.Lotto;
-import lotto.MatchingCountResult;
+import lotto.dto.MatchingCountResult;
+import lotto.model.Lotto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
